refactor: add AnimatableHeaderScrollView and AnimatableHeaderFlatList
