--- src/lib/Components/CollapsibleHeader/CollapsibleHeaderFlatList.tsx
@@ -1,9 +1,0 @@
-import { PropsWithChildren } from "react"
-import { FlatList, FlatListProps, ScrollView } from "react-native"
-import { withCollapsibleHeader } from "./withCollapsibleHeader"
-
-export const CollapsibleHeaderFlatList = withCollapsibleHeader(FlatList) as <T>(
-  props: PropsWithChildren<FlatListProps<T>>
-) => JSX.Element
-
-export const CollapsibleHeaderScrollView = withCollapsibleHeader(ScrollView)
